-Add pragma marks and comments where necessary
