+++ Recommendations/Archiver/RecommendationResultArchiver.swift
@@ -8,23 +8,13 @@
 
 import Foundation
 
+//MARK: RecommendationResultArchiveService
 class RecommendationResultArchiveService: ArchiveService<RecommendationResult> {
     
     static let shared: RecommendationResultArchiveService = RecommendationResultArchiveService(fileName: "recommendationResult.data")
         
     private override init(fileName: String) {
         super.init(fileName: fileName)
-    }
-    
-    override func getAll() -> [RecommendationResult] {
-        let result = super.getAll()
-        print("HERE: get all \(result.count)")
-        return result
-    }
-
-    override func save(_ object: RecommendationResult) {
-        print("HERE: saving ")
-        super.save(object)
     }
 
 }
